Combine native e2e specs (#43844)
