+++ e2e/test/scenarios/native/native-reproductions.cy.spec.ts
@@ -1,4 +1,43 @@
-import { restore, openNativeEditor, runNativeQuery } from "e2e/support/helpers";
+import {
+  restore,
+  withDatabase,
+  adhocQuestionHash,
+  runNativeQuery,
+  openNativeEditor,
+} from "e2e/support/helpers";
+
+describe("issue 11727", { tags: "@external" }, () => {
+  const PG_DB_ID = 2;
+
+  const questionDetails = {
+    dataset_query: {
+      type: "native",
+      database: PG_DB_ID,
+      native: {
+        query: "SELECT pg_sleep(10)",
+      },
+    },
+  };
+  beforeEach(() => {
+    restore("postgres-12");
+    cy.signInAsAdmin();
+    cy.intercept("GET", "/api/database").as("getDatabases");
+  });
+
+  it("should cancel the native query via the keyboard shortcut (metabase#11727)", () => {
+    withDatabase(PG_DB_ID, () => {
+      cy.visit("/question#" + adhocQuestionHash(questionDetails));
+      cy.wait("@getDatabases");
+
+      runNativeQuery({ wait: false });
+      cy.findByText("Doing science...").should("be.visible");
+      cy.get("body").type("{cmd}{enter}");
+      cy.findByText("Here's where your results will appear").should(
+        "be.visible",
+      );
+    });
+  });
+});
 
 describe("issue 16584", () => {
   beforeEach(() => {
