--- e2e/test/scenarios/native/reproductions/20625-prefix-match.cy.spec.js
@@ -1,29 +1,0 @@
-import { restore, openNativeEditor } from "e2e/support/helpers";
-
-describe("issue 20625", { tags: "@quarantine" }, () => {
-  beforeEach(() => {
-    restore();
-    cy.signInAsAdmin();
-    cy.request("PUT", "/api/setting/native-query-autocomplete-match-style", {
-      value: "prefix",
-    });
-    cy.signInAsNormalUser();
-    cy.intercept("GET", "/api/database/*/autocomplete_suggestions**").as(
-      "autocomplete",
-    );
-  });
-
-  // realpress messes with cypress 13
-  it("should continue to request more prefix matches (metabase#20625)", () => {
-    openNativeEditor().type("s");
-
-    // autocomplete_suggestions?prefix=s
-    cy.wait("@autocomplete");
-
-    // can't use cy.type because it does not simulate the bug
-    cy.realPress("o");
-
-    // autocomplete_suggestions?prefix=so
-    cy.wait("@autocomplete");
-  });
-});
